requirements.txt
python-telegram-bot==20.0.0.0a1
oauth2client==4.1.3
google-api-python-client==2.64.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
python-dotenv==0.21.0